+++ CHANGELOG.md
@@ -1324,3 +1324,6 @@ def your_plugin(graph, func, sig):
 def wrapper(ep_info):
     return your_plugin(ep_info.graph, ep_info.func, ep_info.sig)
 ```
+
+
+## version 0.2.16